requirements.txt
# DineSight – deployment-ready dependencies
# Install: pip install -r requirements.txt

fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlmodel>=0.0.14
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
